[CXF-8361] Separating Jakarta EE & Java EE handler chain implementation details
